+++ monument/cli/src/calls.rs
@@ -1,8 +1,7 @@
 use bellframe::{method::LABEL_LEAD_END, PlaceNot, Stage};
-use monument::layout::new::Call;
+use itertools::Itertools;
+use monument::layout::new::{default_calling_positions, Call};
 use serde::{de, Deserialize, Deserializer};
-
-use super::Error;
 
 /// The values of the `base_calls` attribute
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
@@ -20,7 +19,7 @@ impl BaseCalls {
         singles_only: bool,
         bob_weight: Option<f32>,
         single_weight: Option<f32>,
-    ) -> Result<Vec<Call>, Error> {
+    ) -> anyhow::Result<Vec<Call>> {
         // Panic if the comp has less than 4 bells.  I don't expect anyone to use Monument to
         // generate comps on fewer than 8 bells, but we should still check because the alternative
         // is UB
@@ -50,17 +49,21 @@ impl BaseCalls {
         };
 
         if let Some(w) = bob_weight {
-            bob.set_weight(w);
+            bob.weight = w;
         }
         if let Some(w) = single_weight {
-            single.set_weight(w);
+            single.weight = w;
         }
 
         Ok(match (bobs_only, singles_only) {
             (false, false) => vec![bob, single],
             (true, false) => vec![bob],
             (false, true) => vec![single],
-            (true, true) => return Err(Error::BobsOnlyAndSinglesOnly),
+            (true, true) => {
+                return Err(anyhow::Error::msg(
+                    "Can't be both `bobs_only` and `singles_only`",
+                ))
+            }
         })
     }
 }
@@ -96,26 +99,47 @@ pub struct SpecificCall {
     debug_symbol: Option<String>,
     #[serde(default = "lead_end")]
     lead_location: String,
-    calling_positions: Option<Vec<String>>,
+    calling_positions: Option<CallingPositions>,
     #[serde(default = "default_call_score")]
     weight: f32,
 }
 
+/// The different ways the user can specify a set of calling positions
+#[derive(Debug, Clone, Deserialize)]
+#[serde(untagged, deny_unknown_fields)]
+enum CallingPositions {
+    /// The calling positions should be the `char`s in the given string
+    Str(String),
+    /// Each calling position is explicitly listed
+    List(Vec<String>),
+}
+
 impl SpecificCall {
-    pub(crate) fn to_call_spec(&self, stage: Stage) -> Result<Call, Error> {
-        Ok(Call::new(
-            self.symbol.clone(),
-            self.debug_symbol.as_ref().unwrap_or(&self.symbol).clone(),
-            self.calling_positions.clone(),
-            self.lead_location.clone(),
-            PlaceNot::parse(&self.place_notation, stage)
-                .map_err(|e| Error::CallPnParse(self.place_notation.clone(), e))?,
-            self.weight,
-        ))
+    pub(crate) fn to_call_spec(&self, stage: Stage) -> anyhow::Result<Call> {
+        let place_not = PlaceNot::parse(&self.place_notation, stage).map_err(|e| {
+            anyhow::Error::msg(format!(
+                "Can't parse place notation {:?} for call {:?}: {}",
+                self.place_notation,
+                self.debug_symbol.as_ref().unwrap_or(&self.symbol),
+                e
+            ))
+        })?;
+        let calling_positions = match &self.calling_positions {
+            None => default_calling_positions(&place_not),
+            Some(CallingPositions::Str(s)) => s.chars().map(|c| c.to_string()).collect_vec(),
+            Some(CallingPositions::List(positions)) => positions.clone(),
+        };
+        Ok(Call {
+            display_symbol: self.symbol.clone(),
+            debug_symbol: self.debug_symbol.as_ref().unwrap_or(&self.symbol).clone(),
+            calling_positions,
+            lead_location: self.lead_location.clone(),
+            place_not,
+            weight: self.weight,
+        })
     }
 }
 
-#[inline(always)]
 fn lead_end() -> String {
     LABEL_LEAD_END.to_owned()
 }